requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
pillow>=9.5.0
numpy>=1.24.0
tqdm>=4.65.0
accelerate>=0.20.0
safetensors>=0.3.0
# Optional: For IP-Adapter (uncomment if using generate_with_ip_adapter.py)
# ip-adapter>=1.0.0

